parentPanel changes
