+++ src/data/parentData.js
@@ -1,0 +1,50 @@
+export const complaintStats = [
+    { name: 'Total Complaints', value: 12 },
+    { name: 'Resolved', value: 8 },
+    { name: 'Pending', value: 3 },
+    { name: 'In Progress', value: 1 },
+  ];
+  
+  export const statusDistribution = [
+    { name: 'Resolved', value: 8 },
+    { name: 'Pending', value: 3 },
+    { name: 'In Progress', value: 1 },
+  ];
+  
+  export const complaints = [
+    {
+      id: 1,
+      complaintNo: 'CMP2023001',
+      type: 'Academic',
+      description: 'Issue with course material access',
+      status: 'In Progress',
+      date: '2023-01-15',
+      response: 'We are looking into the issue and will update you soon.',
+    },
+    {
+      id: 2,
+      complaintNo: 'CMP2023002',
+      type: 'Hostel',
+      description: 'Room maintenance issue',
+      status: 'Resolved',
+      date: '2023-01-16',
+      response: 'The maintenance work has been completed.',
+    },
+  ];
+  
+  export const feedbackList = [
+    {
+      id: 1,
+      date: '2023-01-10',
+      rating: 4,
+      comment: 'Good response time and resolution.',
+      response: 'Thank you for your feedback!',
+    },
+    {
+      id: 2,
+      date: '2023-01-15',
+      rating: 5,
+      comment: 'Excellent service and support.',
+      response: 'We appreciate your kind words!',
+    },
+  ];
